Initial commit
This is one huge commits that makes it seem like everything
was done in just one sitting but in reality this is a series
of quite uninteresting commits that won't make any sense.

Future commits will probably be a bit more interesting.

Let's just pretend this is the first, right?